Backendin refaktorointia
