+++ osa8/kirjasto-backend/index.js
@@ -1,13 +1,18 @@
 const { ApolloServer } = require('@apollo/server')
-const { startStandaloneServer } = require('@apollo/server/standalone')
-const { GraphQLError } = require('graphql')
+const { expressMiddleware } = require('@apollo/server/express4')
+const {
+  ApolloServerPluginDrainHttpServer,
+} = require('@apollo/server/plugin/drainHttpServer')
+const { makeExecutableSchema } = require('@graphql-tools/schema')
+const express = require('express')
+const cors = require('cors')
+const http = require('http')
 const jwt = require('jsonwebtoken')
 const mongoose = require('mongoose')
 mongoose.set('strictQuery', false)
-const Author = require('./models/author')
-const Book = require('./models/book')
 const User = require('./models/user')
-
+const typeDefs = require('./schema')
+const resolvers = require('./resolvers')
 require('dotenv').config()
 
 const PORT = process.env.PORT
@@ -23,204 +28,34 @@ mongoose
     console.log('error connection to MongoDB:', error.message)
   })
 
-const typeDefs = `
-  type User {
-    username: String!
-    favoriteGenre: String!
-    id: ID!
-  }
-
-  type Token {
-    value: String!
-  }
-
-  type Author {
-    name: String!
-    born: Int
-    id: ID!
-    bookCount: Int!
-  }
-
-  type Book {
-    title: String!
-    published: Int!
-    author: Author!
-    genres: [String!]!
-    id: ID!
-  }
-
-  type Query {
-    me: User
-    authorCount: Int!
-    bookCount: Int!
-    allBooks(author: String, genre: String): [Book!]!
-    allAuthors: [Author!]!
-  }
-
-  type Mutation {
-    createUser(
-      username: String!
-      favoriteGenre: String!
-    ): User
-    login(
-      username: String!
-      password: String!
-    ): Token
-    addBook(
-      title: String!
-      author: String!
-      published: Int!
-      genres: [String!]!
-    ): Book!
-    editAuthor(
-      name: String!
-      setBornTo: Int!
-    ): Author
-  }
-`
-
-const resolvers = {
-  Author: {
-    bookCount: async (root) => {
-      const books = await Book.find({ author: root.id })
-      return books.length
-    },
-  },
-  Book: {
-    author: async (root) => {
-      const author = await Author.findById(root.author)
-      return author
-    },
-  },
-  Query: {
-    me: (root, args, context) => {
-      return context.currentUser
-    },
-    authorCount: async () => Author.collection.countDocuments(),
-    bookCount: async () => Book.collection.countDocuments(),
-    allBooks: async (root, args) => {
-      if (args.genre) {
-        return Book.find({ genres: { $all: [args.genre] } })
-      } else {
-        return Book.find({})
-      }
-    },
-    allAuthors: async (root, args) => {
-      return Author.find({})
-    },
-  },
-  Mutation: {
-    createUser: async (root, args) => {
-      const user = new User({
-        username: args.username,
-        favoriteGenre: args.favoriteGenre,
-      })
-
-      return user.save().catch((error) => {
-        throw new GraphQLError('Creating the user failed', {
-          extensions: {
-            code: 'BAD_USER_INPUT',
-            invalidArgs: args.username,
-            error,
-          },
-        })
-      })
-    },
-    login: async (root, args) => {
-      const user = await User.findOne({ username: args.username })
-
-      if (!user || args.password !== 'secret') {
-        throw new GraphQLError('wrong credentials', {
-          extensions: {
-            code: 'BAD_USER_INPUT',
-          },
-        })
-      }
-
-      const userForToken = {
-        username: user.username,
-        id: user._id,
-      }
-
-      return { value: jwt.sign(userForToken, process.env.JWT_SECRET) }
-    },
-    addBook: async (root, args, { currentUser }) => {
-      if (!currentUser) {
-        throw new GraphQLError('wrong credentials', {
-          extensions: { code: 'BAD_USER_INPUT' },
-        })
-      }
-      const book = new Book({ ...args })
-      const author = await Author.findOne({ name: args.author })
-      if (!author) {
-        try {
-          const newAuthor = new Author({ name: args.author, born: null })
-          await newAuthor.save()
-          book.author = newAuthor
-        } catch (error) {
-          throw new GraphQLError('Saving new author failed', {
-            extensions: {
-              code: 'BAD_USER_INPUT',
-              invalidArgs: args.author,
-              error,
-            },
-          })
+const start = async () => {
+  const app = express()
+  const httpServer = http.createServer(app)
+  const server = new ApolloServer({
+    schema: makeExecutableSchema({ typeDefs, resolvers }),
+    plugins: [ApolloServerPluginDrainHttpServer({ httpServer })],
+  })
+  await server.start()
+  app.use(
+    '/',
+    cors(),
+    express.json(),
+    expressMiddleware(server, {
+      context: async ({ req }) => {
+        const auth = req ? req.headers.authorization : null
+        if (auth && auth.startsWith('Bearer ')) {
+          const decodedToken = jwt.verify(
+            auth.substring(7),
+            process.env.JWT_SECRET
+          )
+          const currentUser = await User.findById(decodedToken.id)
+          return { currentUser }
         }
-      } else {
-        book.author = author
-      }
-      try {
-        await book.save()
-      } catch (error) {
-        throw new GraphQLError('Saving new book failed', {
-          extensions: {
-            code: 'BAD_USER_INPUT',
-            invalidArgs: args.title,
-            error,
-          },
-        })
-      }
-      return book
-    },
-    editAuthor: async (root, args, { currentUser }) => {
-      if (!currentUser) {
-        throw new GraphQLError('wrong credentials', {
-          extensions: { code: 'BAD_USER_INPUT' },
-        })
-      }
-      const author = await Author.findOne({ name: args.name })
-      try {
-        author.born = args.setBornTo
-        author.save()
-      } catch (error) {
-        throw new GraphQLError('Editing author failed', {
-          extensions: {
-            code: 'BAD_USER_INPUT',
-            invalidArgs: [author.name, author.born],
-            error,
-          },
-        })
-      }
-      return author
-    },
-  },
+      },
+    })
+  )
+  httpServer.listen(PORT, () =>
+    console.log(`Server is now running on http://localhost:${PORT}`)
+  )
 }
-
-const server = new ApolloServer({
-  typeDefs,
-  resolvers,
-})
-
-startStandaloneServer(server, {
-  listen: { port: PORT },
-  context: async ({ req, res }) => {
-    const auth = req ? req.headers.authorization : null
-    if (auth && auth.startsWith('Bearer ')) {
-      const decodedToken = jwt.verify(auth.substring(7), process.env.JWT_SECRET)
-      const currentUser = await User.findById(decodedToken.id)
-      return { currentUser }
-    }
-  },
-}).then(({ url }) => {
-  console.log(`Server ready at ${url}`)
-})
+start()
