+++ osa8/kirjasto-backend/resolvers.js
@@ -1,0 +1,134 @@
+const { GraphQLError } = require('graphql')
+const jwt = require('jsonwebtoken')
+const Author = require('./models/author')
+const Book = require('./models/book')
+const User = require('./models/user')
+
+const resolvers = {
+  Author: {
+    bookCount: async (root) => {
+      const books = await Book.find({ author: root.id })
+      return books.length
+    },
+  },
+  Book: {
+    author: async (root) => {
+      const author = await Author.findById(root.author)
+      return author
+    },
+  },
+  Query: {
+    me: (root, args, context) => {
+      return context.currentUser
+    },
+    authorCount: async () => Author.collection.countDocuments(),
+    bookCount: async () => Book.collection.countDocuments(),
+    allBooks: async (root, args) => {
+      if (args.genre) {
+        return Book.find({ genres: { $all: [args.genre] } })
+      } else {
+        return Book.find({})
+      }
+    },
+    allAuthors: async (root, args) => {
+      return Author.find({})
+    },
+  },
+  Mutation: {
+    createUser: async (root, args) => {
+      const user = new User({
+        username: args.username,
+        favoriteGenre: args.favoriteGenre,
+      })
+
+      return user.save().catch((error) => {
+        throw new GraphQLError('Creating the user failed', {
+          extensions: {
+            code: 'BAD_USER_INPUT',
+            invalidArgs: args.username,
+            error,
+          },
+        })
+      })
+    },
+    login: async (root, args) => {
+      const user = await User.findOne({ username: args.username })
+
+      if (!user || args.password !== 'secret') {
+        throw new GraphQLError('wrong credentials', {
+          extensions: {
+            code: 'BAD_USER_INPUT',
+          },
+        })
+      }
+
+      const userForToken = {
+        username: user.username,
+        id: user._id,
+      }
+
+      return { value: jwt.sign(userForToken, process.env.JWT_SECRET) }
+    },
+    addBook: async (root, args, { currentUser }) => {
+      if (!currentUser) {
+        throw new GraphQLError('wrong credentials', {
+          extensions: { code: 'BAD_USER_INPUT' },
+        })
+      }
+      const book = new Book({ ...args })
+      const author = await Author.findOne({ name: args.author })
+      if (!author) {
+        try {
+          const newAuthor = new Author({ name: args.author, born: null })
+          await newAuthor.save()
+          book.author = newAuthor
+        } catch (error) {
+          throw new GraphQLError('Saving new author failed', {
+            extensions: {
+              code: 'BAD_USER_INPUT',
+              invalidArgs: args.author,
+              error,
+            },
+          })
+        }
+      } else {
+        book.author = author
+      }
+      try {
+        await book.save()
+      } catch (error) {
+        throw new GraphQLError('Saving new book failed', {
+          extensions: {
+            code: 'BAD_USER_INPUT',
+            invalidArgs: args.title,
+            error,
+          },
+        })
+      }
+      return book
+    },
+    editAuthor: async (root, args, { currentUser }) => {
+      if (!currentUser) {
+        throw new GraphQLError('wrong credentials', {
+          extensions: { code: 'BAD_USER_INPUT' },
+        })
+      }
+      const author = await Author.findOne({ name: args.name })
+      try {
+        author.born = args.setBornTo
+        author.save()
+      } catch (error) {
+        throw new GraphQLError('Editing author failed', {
+          extensions: {
+            code: 'BAD_USER_INPUT',
+            invalidArgs: [author.name, author.born],
+            error,
+          },
+        })
+      }
+      return author
+    },
+  },
+}
+
+module.exports = resolvers
